+++ generators/go/internal/testdata/model/union/fixtures/types.go
@@ -11,6 +11,29 @@ import (
 
 type Bar struct {
 	Name string `json:"name" url:"name"`
+
+	extraProperties map[string]interface{}
+}
+
+func (b *Bar) GetExtraProperties() map[string]interface{} {
+	return b.extraProperties
+}
+
+func (b *Bar) UnmarshalJSON(data []byte) error {
+	type unmarshaler Bar
+	var value unmarshaler
+	if err := json.Unmarshal(data, &value); err != nil {
+		return err
+	}
+	*b = Bar(value)
+
+	extraProperties, err := core.ExtractExtraProperties(data, *b)
+	if err != nil {
+		return err
+	}
+	b.extraProperties = extraProperties
+
+	return nil
 }
 
 func (b *Bar) String() string {
@@ -22,6 +45,12 @@ func (b *Bar) String() string {
 
 type Baz struct {
 	extended string
+
+	extraProperties map[string]interface{}
+}
+
+func (b *Baz) GetExtraProperties() map[string]interface{} {
+	return b.extraProperties
 }
 
 func (b *Baz) Extended() string {
@@ -40,6 +69,13 @@ func (b *Baz) UnmarshalJSON(data []byte) error {
 	}
 	*b = Baz(unmarshaler.embed)
 	b.extended = "extended"
+
+	extraProperties, err := core.ExtractExtraProperties(data, *b, "extended")
+	if err != nil {
+		return err
+	}
+	b.extraProperties = extraProperties
+
 	return nil
 }
 
@@ -64,6 +100,29 @@ func (b *Baz) String() string {
 
 type Foo struct {
 	Name string `json:"name" url:"name"`
+
+	extraProperties map[string]interface{}
+}
+
+func (f *Foo) GetExtraProperties() map[string]interface{} {
+	return f.extraProperties
+}
+
+func (f *Foo) UnmarshalJSON(data []byte) error {
+	type unmarshaler Foo
+	var value unmarshaler
+	if err := json.Unmarshal(data, &value); err != nil {
+		return err
+	}
+	*f = Foo(value)
+
+	extraProperties, err := core.ExtractExtraProperties(data, *f)
+	if err != nil {
+		return err
+	}
+	f.extraProperties = extraProperties
+
+	return nil
 }
 
 func (f *Foo) String() string {