+++ generators/python/src/fern_python/generators/fastapi/fastapi_generator.py
@@ -41,7 +41,7 @@ class FastApiGenerator(AbstractGenerator):
     ) -> Tuple[str, ...]:
         return (
             generator_config.organization,
-            ir.api_name.snake_case.unsafe_name,
+            ir.api_name.snake_case.safe_name,
         )
 
     def run(