+++ seed/fastapi/reserved-keywords/resources/package/service/service.py
@@ -17,21 +17,25 @@ from ....core.route_args import get_route_args
 class AbstractPackageService(AbstractFernService):
     """
     AbstractPackageService is an abstract class containing the methods that you should implement.
-    
+
     Each method is associated with an API route, which will be registered
     with FastAPI when you register your implementation using Fern's register()
     function.
     """
+
     @abc.abstractmethod
-    def test(self, *, for: str) -> None:
+    def test(self, *, for_: str) -> None:
         ...
+
     """
     Below are internal methods used by Fern to register your implementation.
     You can ignore them.
     """
+
     @classmethod
     def _init_fern(cls, router: fastapi.APIRouter) -> None:
         cls.__init_test(router=router)
+
     @classmethod
     def __init_test(cls, router: fastapi.APIRouter) -> None:
         endpoint_function = inspect.signature(cls.test)
@@ -39,12 +43,12 @@ class AbstractPackageService(AbstractFernService):
         for index, (parameter_name, parameter) in enumerate(endpoint_function.parameters.items()):
             if index == 0:
                 new_parameters.append(parameter.replace(default=fastapi.Depends(cls)))
-            elif parameter_name == "for":
-                new_parameters.append(parameter.replace(default=fastapi.Query(default=...)))
+            elif parameter_name == "for_":
+                new_parameters.append(parameter.replace(default=fastapi.Query(default=..., alias="for")))
             else:
                 new_parameters.append(parameter)
         setattr(cls.test, "__signature__", endpoint_function.replace(parameters=new_parameters))
-        
+
         @functools.wraps(cls.test)
         def wrapper(*args: typing.Any, **kwargs: typing.Any) -> None:
             try:
@@ -56,11 +60,11 @@ class AbstractPackageService(AbstractFernService):
                     + "the endpoint's errors list in your Fern Definition."
                 )
                 raise e
-        
+
         # this is necessary for FastAPI to find forward-ref'ed type hints.
         # https://github.com/tiangolo/fastapi/pull/5077
         wrapper.__globals__.update(cls.test.__globals__)
-        
+
         router.post(
             path="/",
             response_model=None,