+++ seed/go-model/objects-with-imports/commons/types.go
@@ -3,6 +3,7 @@
 package commons
 
 import (
+	json "encoding/json"
 	fmt "fmt"
 	core "github.com/objects-with-imports/fern/core"
 )
@@ -10,6 +11,29 @@ import (
 type Metadata struct {
 	Id   string            `json:"id" url:"id"`
 	Data map[string]string `json:"data,omitempty" url:"data,omitempty"`
+
+	extraProperties map[string]interface{}
+}
+
+func (m *Metadata) GetExtraProperties() map[string]interface{} {
+	return m.extraProperties
+}
+
+func (m *Metadata) UnmarshalJSON(data []byte) error {
+	type unmarshaler Metadata
+	var value unmarshaler
+	if err := json.Unmarshal(data, &value); err != nil {
+		return err
+	}
+	*m = Metadata(value)
+
+	extraProperties, err := core.ExtractExtraProperties(data, *m)
+	if err != nil {
+		return err
+	}
+	m.extraProperties = extraProperties
+
+	return nil
 }
 
 func (m *Metadata) String() string {